Make Users list subscriptions via Apollo component
