+++ packages/client/src/modules/user/containers/UserOperations.js
@@ -1,5 +1,4 @@
 import { graphql } from 'react-apollo';
-import update from 'immutability-helper';
 import { removeTypename } from '../../../../../common/utils';
 
 import USERS_STATE_QUERY from '../graphql/UsersStateQuery.client.graphql';
@@ -7,7 +6,6 @@ import UPDATE_ORDER_BY from '../graphql/UpdateOrderBy.client.graphql';
 import USERS_QUERY from '../graphql/UsersQuery.graphql';
 import DELETE_USER from '../graphql/DeleteUser.graphql';
 import UPDATE_FILTER from '../graphql/UpdateFilter.client.graphql';
-import USERS_SUBSCRIPTION from '../graphql/UsersSubscription.graphql';
 
 const withUsersState = Component =>
   graphql(USERS_STATE_QUERY, {
@@ -74,54 +72,4 @@ const withFilterUpdating = Component =>
     })
   })(Component);
 
-function addUser(prev, node) {
-  return update(prev, {
-    users: {
-      $set: [...prev.users, node]
-    }
-  });
-}
-
-function deleteUser(prev, id) {
-  const index = prev.users.findIndex(user => user.id === id);
-  // ignore if not found
-  if (index < 0) {
-    return prev;
-  }
-  return update(prev, {
-    users: {
-      $splice: [[index, 1]]
-    }
-  });
-}
-
-const subscribeToUsersList = (subscribeToMore, filter) => {
-  return subscribeToMore({
-    document: USERS_SUBSCRIPTION,
-    variables: { filter },
-    updateQuery: (
-      prev,
-      {
-        subscriptionData: {
-          data: {
-            usersUpdated: { mutation, node }
-          }
-        }
-      }
-    ) => {
-      switch (mutation) {
-        case 'CREATED':
-          return addUser(prev, node);
-        case 'DELETED':
-          return deleteUser(prev, node.id);
-        case 'UPDATED':
-          return deleteUser(prev, node.id);
-        default:
-          return prev;
-      }
-    }
-  });
-};
-
 export { withUsersState, withUsers, withUsersDeleting, withOrderByUpdating, withFilterUpdating };
-export { subscribeToUsersList };
